sample 1/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1
python-Levenshtein==0.21.1
sentence-transformers==2.2.2
openpyxl==3.1.2
